refactor Handles to new file
