+++ pol-core/pol/CMakeSources.cmake
@@ -1,5 +1,5 @@
 set (pol_sources  # sorted !
-  CMakeSources.cmake 
+  CMakeSources.cmake
   PolMain.cpp
   PolMain.h
   StdAfx.h
@@ -59,6 +59,8 @@ set (pol_sources  # sorted !
   crypt/twofish.h
   ctable.cpp
   ctable.h
+  dap/handles.cpp
+  dap/handles.h
   dap/server.cpp
   dap/server.h
   dblclick.cpp
